pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snapcal"
version = "0.1.0"
description = "SAM-guided food recognition and calorie estimation pipeline."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.26,<3.0",
  "Pillow>=10.4,<11.0",
]

[project.optional-dependencies]
api = [
  "Django>=4.2,<5.0",
  "djangorestframework>=3.15,<4.0",
  "django-cors-headers>=4.4,<5.0",
  "dj-database-url>=2.2,<3.0",
  "gunicorn>=23.0,<24.0",
  "mobile-sam>=1.0.0",
  "psycopg2-binary>=2.9,<3.0",
  "torch>=2.4,<3.0",
  "torchvision>=0.19,<1.0",
  "transformers>=4.45,<5.0",
  "uvicorn>=0.35,<1.0",
  "whitenoise>=6.7,<7.0",
]
train = [
  "matplotlib>=3.8,<4.0",
  "mobile-sam>=1.0.0",
  "pandas>=2.2,<3.0",
  "scikit-learn>=1.5,<2.0",
  "torch>=2.4,<3.0",
  "torchvision>=0.19,<1.0",
  "transformers>=4.45,<5.0",
]
dev = [
  "pytest>=8.3,<9.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
